requirements.txt
python-telegram-bot
python-dotenv
pymediainfo
aiohttp
google-generativeai